Generate secure blocks inline... makes for a cleaner experience.

+++ tools/postprocess.sh
@@ -9,9 +9,11 @@ source private.env
 for pf in `find jacoblewallen.com/public/albums/ -iname "*.html"`; do
 	echo $pf
 
-	build/secure --passphrase ${PASSPHRASE} --plaintext $pf --ciphertext $pf.aes
+	build/secure --inline --passphrase ${PASSPHRASE} --plaintext $pf --ciphertext $pf.aes
 
-	mv $pf.aes $pf
+	if [ -f $pf.aes ]; then
+		mv $pf.aes $pf
+	fi
 done
 
 popd
